pyproject.toml
[tool.poetry]
name = "keke-bot"
version = "0.1.0"
description = "keke-bot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_manager", "nonebot_plugin_heisi", "random_cat_gif", "nonebot_plugin_withdraw", "nonebot_plugin_status", "nonebot_plugin_crazy_thursday", "nonebot_plugin_alias"]
plugin_dirs = ["keke_bot/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
